feat(tool): 添加Base64图像提取功能
- 新增 base64_utils.hpp 工具类，支持Base64编解码和图像类型检测
- 新增 test_image_extraction.cpp 测试程序
- 在 maxwell_parser.hpp 中添加 extractBase64Image 接口
- 成功从 Project49.aedt 提取 200x200 JPEG 图像

同时修复编译器警告:
- 添加 (void) 抑制未使用参数警告
- 添加缺失的 #include <dirent.h>
- 移除 preconditioner.cpp 中重复的模板实例化

+++ src/tool/maxwell_parser.cpp
@@ -105,7 +105,7 @@ void MaxwellParser::preprocess_content() {
 }
 
 void MaxwellParser::skip_whitespace_and_comments() {
-    while (current_line_ < lines_.size()) {
+    while (current_line_ < static_cast<int>(lines_.size())) {
         std::string line = lines_[current_line_];
         
         // 跳过空白行
@@ -127,7 +127,7 @@ void MaxwellParser::skip_whitespace_and_comments() {
 std::shared_ptr<BlockNode> MaxwellParser::parse_block() {
     skip_whitespace_and_comments();
     
-    if (current_line_ >= lines_.size()) {
+    if (current_line_ >= static_cast<int>(lines_.size())) {
         return nullptr;
     }
     
@@ -143,11 +143,10 @@ std::shared_ptr<BlockNode> MaxwellParser::parse_block() {
     auto block = std::make_shared<BlockNode>(block_name, current_line_ + 1);
     current_line_++;
     
-    // 解析块内容
-    while (current_line_ < lines_.size()) {
+    while (current_line_ < static_cast<int>(lines_.size())) {
         skip_whitespace_and_comments();
         
-        if (current_line_ >= lines_.size()) {
+        if (current_line_ >= static_cast<int>(lines_.size())) {
             throw ParseError("块未正确结束: " + block_name, block->start_line);
         }
         
@@ -192,7 +191,7 @@ std::shared_ptr<BlockNode> MaxwellParser::parse_block() {
 }
 
 Property MaxwellParser::parse_property() {
-    if (current_line_ >= lines_.size()) {
+    if (current_line_ >= static_cast<int>(lines_.size())) {
         throw ParseError("期望属性定义", current_line_ + 1);
     }
